new: Cast NSArray objects to their target type
When retrieving objects from an NSArray, we don't want to get a bunch of
NSObjects and have to deal with casting ourselves.
The solution to this is to create a new instance of the desired class
when fetching an object from the array.

We need some better error handling in relation to this, right now it'll
just print a stacktrace and return null. Should we print a stacktrace
at all?